testing mysql support
