+++ src/main/java/cat/nyaa/nyaacore/database/DataColumn.java
@@ -9,4 +9,5 @@ import java.lang.annotation.Target;
 @Retention(RetentionPolicy.RUNTIME)
 public @interface DataColumn {
     String value() default "";
+    int length() default 0;
 }
